+++ app/models/move.rb
@@ -14,6 +14,7 @@ class Move < ApplicationRecord
   belongs_to :user
   has_many :move_tags
   has_many :tags, through: :move_tags
+  has_many :user_moves, -> { order(:practiced_on) }
 
   # Methods
 
@@ -27,5 +28,13 @@ class Move < ApplicationRecord
     clean_tokens.each do |token|
       tags << Tag.find_or_create_tag(token, user_id)
     end
+  end
+
+  def add_practice_date(current_user, date)
+    user_moves.where(user_id: current_user.id, practiced_on: date).first_or_create
+  end
+
+  def practiced_today?
+    user_moves.where(practiced_on: Time.zone.today).count == 1
   end
 end