+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - **Moves Added**
   - Added ability to create, view, modify, and search through moves
   - Moves can be filtered by clicking on tag names on the moves index
+  - Added ability to indicate that the move was practiced today
 - **Tags Added**
   - Moves can be tagged
   - Tags can be used to search for similar moves